Exercise 2.5. The Power Of One Write a program that does the following ten times (so, ten full loops with the time computed): Randomly select a single, lowercase letter. This is the “preimage seed.”
Use MD5 to compute the hash of this initial letter. This is the “test hash.”

In a loop, iterate through all possible lowercase one-letter inputs.
Hash each letter in the same way as before, and compare against the test hash.
When you find a match, stop.
Compute the amount of time it took to find a match.

How long, on average, did it take to find a match for a random preimage seed?

Exercise 2.6. The Power of One, But Bigger!
Repeat the previous exercise, but use an increasingly bigger input alphabet set. Try the test with both lowercase and uppercase letters. Then try it with lowercase letters, uppercase letters, and numbers. Finally, try all printable characters (string.printable).
How many total symbols are in each input set?

How much longer does each run take?

Exercise 2.7. Password Length’s Effects on Attack Time
Repeat the previous exercise, but this time for two-symbol inputs. Then try it with three and four symbols at a time. How much longer does it take to invert the randomly chosen input?